inject mock item provider when needed
Determines which provider to use,  mock item provider for screenshot tests , or default PointOfSaleItemService
