+++ WooCommerce/WooCommerceScreenshots/WooCommerceScreenshots.swift
@@ -28,7 +28,7 @@ class WooCommerceScreenshots: XCTestCase {
         app.launchArguments.append("disable-animations")
         app.launchArguments.append("-mocks-push-notification")
         app.launchArguments.append("bypass-pos-eligibility-checks")
-        app.launchArguments.append("bypass-pos-product-loading")
+        app.launchArguments.append("load-mocked-pos-products")
         app.launchArguments.append("bypass-pos-order-syncing")
         app.launchArguments.append("bypass-card-present-payment")
         app.launchArguments.append(contentsOf: ["-mocks-port", "\(server.listenAddress.port)"])
